requirements.txt
numpy>=1.24
torch>=2.0
pygame-ce>=2.5
